videogame and released_for: refactoring and bugfix
The two transformations were simplified (in case we want to include
screenshots in the final report).

Bugfix: the regex normalization over the titles punctuations (both in
videogame and released_for) was not performed immediately at the data
sources, but later on in the integration process. In this way, the
number of released_for tuples was incremented, as well as the amount of
tuples in released_for with non-null sales:
 - tuples: 20'773 -> 25'511
 - tuples with non-null sales: ca. 7'000 -> ca. 12'000.

Signed-off-by: Alessio Papi <alessio.papi@live.it>